LPS-147901 Modularize DXPActivationKeys components to simplify
