+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/containers/DXPActivationKeysTable/hooks/useGetActivationKeysData.js
@@ -46,9 +46,8 @@ export default function useGetActivationKeysData(project, sessionId) {
 	}, [filterTerm, licenseKeyDownloadURL, project?.accountKey, sessionId]);
 
 	return {
-		activationKeys,
+		activationKeysState: [activationKeys, setActivationKeys],
 		loading,
-		setActivationKeys,
 		setFilterTerm,
 	};
 }
